중복 코드 삭제
