+++ app/src/main/java/com/wafflestudio/snutt2/data/notifications/NotificationRepositoryImpl.kt
@@ -7,7 +7,6 @@ import androidx.paging.map
 import com.wafflestudio.snutt2.domain_model.Notification
 import com.wafflestudio.snutt2.domain_model.domainModel
 import com.wafflestudio.snutt2.lib.network.SNUTTRestApi
-import com.wafflestudio.snutt2.lib.network.dto.core.NotificationDto
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
@@ -16,16 +15,6 @@ import javax.inject.Singleton
 @Singleton
 class NotificationRepositoryImpl @Inject constructor(private val api: SNUTTRestApi) :
     NotificationRepository {
-    override suspend fun getNotificationResultStream(): Flow<PagingData<NotificationDto>> {
-        return Pager(
-            config = PagingConfig(
-                pageSize = NOTIFICATIONS_LOAD_PAGE_SIZE,
-                enablePlaceholders = false,
-            ),
-            pagingSourceFactory = { NotificationPagingSource(api) },
-        ).flow
-    }
-
     override suspend fun getNotificationListStream(): Flow<PagingData<Notification>> {
         return Pager(
             config = PagingConfig(
