+++ app/src/main/java/com/wafflestudio/snutt2/views/RootActivity.kt
@@ -72,7 +72,7 @@ import com.wafflestudio.snutt2.views.logged_in.lecture_detail.LectureColorSelect
 import com.wafflestudio.snutt2.views.logged_in.lecture_detail.LectureDetailPage
 import com.wafflestudio.snutt2.views.logged_in.lecture_detail.LectureDetailViewModel
 import com.wafflestudio.snutt2.views.logged_in.lecture_detail.deeplink.TimetableLectureDetailPage
-import com.wafflestudio.snutt2.views.logged_in.notifications.NotificationPage
+import com.wafflestudio.snutt2.views.logged_in.notifications.NotificationRoute
 import com.wafflestudio.snutt2.views.logged_in.table_lectures.LecturesOfTablePage
 import com.wafflestudio.snutt2.views.logged_in.vacancy_noti.VacancyPage
 import com.wafflestudio.snutt2.views.logged_in.vacancy_noti.VacancyViewModel
@@ -258,7 +258,7 @@ class RootActivity : AppCompatActivity() {
 
                     composable2(NavigationDestination.ImportantNotice) { ImportantNoticePage() }
 
-                    composable2(NavigationDestination.Notification) { NotificationPage() }
+                    composable2(NavigationDestination.Notification) { NotificationRoute() }
 
                     composable2(NavigationDestination.LecturesOfTable) { LecturesOfTablePage() }
